Factor macros generating code for each DataLayout POD type
Summary:
In multiple places, we need to generate template code for each POD type supported, which can be error prone. This System makes it easy to repeat an existing preprocessor macro for each POD type supported by DataLayout.
Having this shared definition helps guarantee that we can repeat the behaviors everywhere with the same list of POD type defined here.

Reviewed By: kiminoue7

Differential Revision: D45020691

fbshipit-source-id: baa96ddde808f43b1f52cc9f4d7abc4966f2a9c9