Fix Windows builds -> ssize_t is a posix only type
Summary:
https://pubs.opengroup.org/onlinepubs/9699919799/basedefs/sys_types.h.html

ssize_t is only a part of the posix standard and seems to be equivalent to size_t but with the caveat it should also be able to store -1? Seems we should be using the python define e.g. `py::ssize_t` so let's use that (before it was working "by accident")

Reviewed By: kiminoue7

Differential Revision: D47670541

fbshipit-source-id: b57a4e5b17d05a15e27030aac76edd8409864afb