Fix AsyncVRSReader's crash
Summary:
AsyncVRSReader sometimes crashed during the read and it came from either at GC time where the reference is unavailable when it tries to update the ref count or when trying to set the future in C++ loop.
Considering that we need to hold the GIL when creating an object to avoid race, adding GIL around the code where we obtain loop, future from the Python runtime.

Reviewed By: georges-berenger

Differential Revision: D55805967

fbshipit-source-id: 888506d23c57f25b91e561ec79307931f025ef26