Add API to get each stream's disk size (#141)
Summary:

This diff provides easy methods to know how much disk space is used by a stream. The API is separate from existing APIs, because the computation is fairly expensive and might require non-trivial memory allocation to track record boundaries, and we don't want to pay for this when not necessary.

Reviewed By: paulsammut

Differential Revision: D67491139