Add API to get each stream's disk size (#141)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pyvrs/pull/141

This diff provides easy methods to know how much disk space is used by a stream. The API is separate from existing APIs, because the computation is fairly expensive and might require non-trivial memory allocation to track record boundaries, and we don't want to pay for this when not necessary.

Reviewed By: paulsammut

Differential Revision: D67491139

fbshipit-source-id: 0b1bd6fc4b8ce841e99f5e060ee172aa5497c26f