Fix AsyncVRSReader crasher
Summary:
When we call:

workerQueue_.waitForJob(job, 1)

because job might be still be set to the last job executed, and it might be deleted in waitForJob(). The problem is that AsyncJob contains two py::object that can be deleted, too, but we're not holding the GIL at that point, so the delete can lead to a GIL violation and a crash.

Reviewed By: jtbraun

Differential Revision: D67784827

fbshipit-source-id: 9dfb2c871b33ca2dc168fa3ebc0808c6ec42bb7d