Fix bindings stubs generation
Summary:
Regenerated bindings

```
buck run arvr/mode/platform010/opt //arvr/python/pyvrs2:vrsbindings_stubgen
```

This required a decent amount of work, because pybindings actually expect things to be declared in order of usage. In other words, you shouldn't declare a pybind for  `void do_something(MyType foo);` before the pybinding `MyType` itself. 

This makes sense, but in practice, the VRS pybindings were breaking these rules in a myriad of ways, so I had to move a lot of things around, and it was quite annoying to do so...

There is a way to prevent regression for this, I will reach out to VRS team about this...

Differential Revision: D69120247