Fix bindings stubs generation (#153)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pyvrs/pull/153

Regenerated bindings

This required a decent amount of work, because pybindings actually expect things to be declared in order of usage. In other words, you shouldn't declare a pybind for  `void do_something(MyType foo);` before the pybinding `MyType` itself.

This makes sense, but in practice, the VRS pybindings were breaking these rules in a myriad of ways, so I had to move various pybind declarations around

There is a way to prevent regression for this, I will work on this separatel

Reviewed By: georges-berenger

Differential Revision: D69120247

fbshipit-source-id: 7697f8021c088319c9faba78698f0865e38ad98b