Rename "Buffer" to "ContentBuffer"
Summary:
This is a flaw in pyvrs2: it uses "Buffer" as the name for a content buffer. Unfortunately the name conflicts with the Buffer type annotation, and ends up creating broken python stubs

I think the only sane fix forward is to just not name this "Buffer", unfortunately. I understand this is a breaking change, but I really don't have another good way forward here....

Reviewed By: georges-berenger

Differential Revision: D69143136

fbshipit-source-id: fa5af362926240a9f3ffd4e590b76d414dfd2f73