Make it possible to access metadata when processing an image
Summary:
When filtering a VRS using AsyncImageFilter, the metadata part of the record is only available when the record is read, which is a problem when trying to process records in parallel from different threads.
With this diff, you can ask tell the filter you'll need metadata, then request an image's metadata when you have it for processing. Note that you can not ask for any metadata, only that of the record you're processing.

Reviewed By: paulsammut

Differential Revision: D74054976

fbshipit-source-id: 75591ae0a9a0d31baa88a6fde850982117a453a8